Update: Move TemplateAreas to editor.
