+++ packages/edit-site/src/components/sidebar-edit-mode/template-panel/style.scss
@@ -1,25 +1,4 @@
 .edit-site-template-card {
-	&__template-areas {
-		margin-top: $grid-unit-20;
-	}
-
-	&__template-areas-list {
-		margin: 0;
-
-		> li {
-			margin: 0;
-		}
-	}
-
-	&__template-areas-item {
-		width: 100%;
-
-		// Override the default padding.
-		&.components-button.has-icon {
-			padding: 0;
-		}
-	}
-
 	&__actions {
 		line-height: 0;
 		> .components-button.is-small.has-icon {
@@ -30,14 +9,6 @@
 	}
 }
 
-h3.edit-site-template-card__template-areas-title {
-	font-weight: 500;
-	margin: 0 0 $grid-unit-10;
-}
-
-.edit-site-template-card__templates-list {
-	margin-top: $grid-unit-20;
-}
 
 .edit-site-template-panel .block-editor-block-preview__container {
 	border-radius: 2px;
