+++ packages/editor/src/components/template-areas/style.scss
@@ -1,0 +1,24 @@
+.editor-template-areas {
+    margin-top: $grid-unit-20;
+    &__list {
+		margin-top: $grid-unit-20;
+
+		> li {
+			margin: 0;
+		}
+	}
+
+    &__item {
+		width: 100%;
+
+		// Override the default padding.
+		&.components-button.has-icon {
+			padding: 0;
+		}
+	}
+}
+
+h3.editor-template-areas__title {
+	font-weight: 500;
+	margin: 0 0 $grid-unit-10;
+}
