+++ packages/editor/src/style.scss
@@ -30,4 +30,5 @@
 @import "./components/post-trash/style.scss";
 @import "./components/preview-dropdown/style.scss";
 @import "./components/table-of-contents/style.scss";
+@import "./components/template-areas/style.scss";
 @import "./components/template-validation-notice/style.scss";
